MH-10165, Empty Dictionary Results may Cause Errors
This commit fixes the problem that an empty result from a
DictionaryService will raise an exception as the TextualImpl class will
not accept empty strings.

This fix will now check if the result is empty and return null instead.
The TextAnalyzerService will check the result and only include actual
values. The positive side-effect of this is, that empty strings may not
show up in the resulting mpeg7 catalogs anymore.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>